+++ src/critical.rs
@@ -1,4 +1,4 @@
-pub const CRITICAL_MULTIPLIER: u8 = 2;
+pub const CRITICAL_MULTIPLIER: u16 = 2;
 const CRITICAL_PERCENT: f32 = 0.1;
 
 pub struct Critical {
